# openapi.yml
openapi: 3.0.0
info:
  title: ASAP API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/auth/login-link:
    post:
      summary: Create login link
      operationId: createLoginLink
      tags:
        - auth
      description: Generate a login link that will be sent to the user's email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginLinkRequest"
      responses:
        "201":
          description: Login link created successfully
        "400":
          description: Invalid email format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/exchange:
    get:
      summary: Exchange login link token for JWT
      operationId: exchangeLoginLink
      tags:
        - auth
      description: Exchange a login link token for a JWT access token by clicking the link from email
      parameters:
        - name: organizationId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Optional organization ID if the login link is associated with a specific organization
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: The login link token received via email
        - name: twoFactorAuthCode
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Token exchange successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMultiResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/refresh:
    post:
      summary: Refresh JWT token
      operationId: refreshToken
      tags:
        - auth
      description: Refresh an existing JWT token to extend its lifetime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/login:
    post:
      summary: Login with email and password
      operationId: loginUser
      tags:
        - auth
      description: Authenticate user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMultiResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/forgot-password:
    post:
      summary: Request password reset
      operationId: forgotPassword
      tags:
        - auth
      description: Send password reset link to user's email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Password reset email sent
        "400":
          description: Invalid email format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/reset-password:
    post:
      summary: Reset password with token
      operationId: resetPassword
      tags:
        - auth
      description: Reset user password using reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetPasswordMultiResponse"
        "400":
          description: Invalid request or token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ========= Users =========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginLinkRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Email address of the user to send the login link to
      required:
        - email
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: JWT refresh token
      required:
        - refreshToken
    LoginRequest:
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
          description: Organization ID to login to
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          description: User's password
        twoFactorAuthCode:
          type: string
      required:
        - email
        - password
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
      required:
        - email
    AuthMultiResponse:
      oneOf:
        - $ref: "#/components/schemas/AuthResponse"
        - $ref: "#/components/schemas/AvailableOrganizationsResponse"
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
          description: Password reset token
        newPassword:
          type: string
          minLength: 8
          description: New password (minimum 8 characters)
      required:
        - token
        - newPassword
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        twoFactorAuthToken:
          type: string
          description: Temporary token for 2FA verification (if applicable)
        twoFactorAuthRequired:
          type: boolean
    ResetPasswordResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the password reset was successful
    ResetPasswordMultiResponse:
      oneOf:
        - $ref: "#/components/schemas/AvailableOrganizationsResponse"
        - $ref: "#/components/schemas/ResetPasswordResponse"
...
    Error:
      type: object
      properties:
        code:
          type: integer
          description: Error code
        message:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
        path:
          type: string
          description: Request path that caused the error
...
    AvailableOrganizationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Organization ID
        name:
          type: string
          description: Organization name
    AvailableOrganizationsResponse:
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/AvailableOrganizationResponse"
          description: List of available organizations for the user
...